requirements.txt
pandas
numpy
requests
python-dotenv
google-api-python-client

pytrends
pyspark
vaderSentiment
scikit-learn
xgboost
lightgbm
pymongo
fastapi
uvicorn
matplotlib
seaborn
plotly
joblib
transformers
torch
nltk
pyarrow